school_report/requirements-new.txt
asgiref==3.8.1
Django==5.2
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
python-dotenv>=1.0.0
django-debug-toolbar>=4.2.0
faker>=19.0.0
django-extensions

# PDF generation
weasyprint>=62.0

# for render.com
whitenoise>=6.5.0

# Production requirements
psycopg2-binary>=2.9.9  # PostgreSQL adapter
gunicorn>=21.2.0  # WSGI HTTP Server
django-storages>=1.14.2  # Storage backends
boto3>=1.34.0  # AWS S3 integration
dj-database-url>=2.1.0  # Database URL parsing for cloud deployments

# Activity tracking and audit logging
django-auditlog>=3.0.0  # Complete audit trail for compliance
django-activity-stream>=2.0.0  # User-facing activity feeds
